Various bugfixes
Fix username validation. We now allow numbers.
Fix contents "createdAt" and "updatedAt" validation. The values are set in
a prePersist Doctrine eventListener, so the values are null when the form
validation happens when we create a new content. The Doctrine mapping
doesn't allow null values, so we are safe database-wise.
Fix a bug where the use of medias in SonataClassificationBundle was
broken, due tu a bugfix from their part that makes it require to add some
config to provide the bundle with the Media class.